ENH: Attach arrowheads to edges in SVG graph renderer.
